requirements.txt
# Backend API - Sistema de Gestão de Sugestões
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0

# Firebase Database
firebase-admin>=6.2.0
google-cloud-firestore>=2.13.0

# Google APIs para sincronização
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utils
pydantic>=2.5.0
email-validator>=2.1.0
aiofiles>=23.2.0
jinja2>=3.1.0

# API Testing
requests>=2.31.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
